너덜너덜

+++ src/main/java/himedia/myportal/repositories/BoardDao.java
@@ -7,8 +7,8 @@ import himedia.myportal.repositories.vo.BoardVo;
 public interface BoardDao {
 	public List<BoardVo> selectAll();
 	public int insert(BoardVo boardVo);
-	public BoardVo getContent(Long no);
+	public BoardVo getContent(Integer no);
 	public int update(BoardVo boardVo);
-	public int delete(Long no, Long userNo);
+	public int delete(Integer no, Integer userNo);
 
 }
